file prep before new feature integration
